WT-1911 - refactor constatnt/env location
