+++ packages/checkout/sdk/src/balances/balances.test.ts
@@ -23,7 +23,7 @@ import {
   Blockscout,
   BlockscoutNativeTokenData, BlockscoutToken, BlockscoutTokens, BlockscoutTokenType,
 } from '../client';
-import { BLOCKSCOUT_CHAIN_URL_MAP, ERC20ABI, IMX_ADDRESS_ZKEVM } from '../lib';
+import { BLOCKSCOUT_CHAIN_URL_MAP, ERC20ABI, IMX_ADDRESS_ZKEVM } from '../env';
 
 jest.mock('../tokens');
 jest.mock('../client');
