+++ packages/checkout/sdk/src/fiatRamp/fiatRamp.ts
@@ -3,7 +3,7 @@ import {
   OnRampConfig, OnRampProvider, OnRampProviderFees,
 } from '../types';
 import { CheckoutConfiguration } from '../config';
-import { TRANSAK_API_BASE_URL } from '../lib';
+import { TRANSAK_API_BASE_URL } from '../env';
 
 export interface FiatRampWidgetParams {
   exchangeType: ExchangeType;
