+++ packages/checkout/sdk/src/gasEstimate/gasEstimator.ts
@@ -24,7 +24,7 @@ import {
 } from './bridgeGasEstimate';
 import * as instance from '../instance';
 import { CheckoutConfiguration, getL1ChainId, getL2ChainId } from '../config';
-import { ERC20ABI } from '../lib';
+import { ERC20ABI } from '../env';
 
 const DUMMY_WALLET_ADDRESS = '0x0000000000000000000000000000000000000001';
 const DEFAULT_TOKEN_DECIMALS = 18;
