+++ packages/checkout/sdk/src/network/network.test.ts
@@ -21,7 +21,7 @@ import { CheckoutError, CheckoutErrorType } from '../errors';
 import { CheckoutConfiguration } from '../config';
 import { RemoteConfigFetcher } from '../config/remoteConfigFetcher';
 import { getUnderlyingChainId } from '../provider/getUnderlyingProvider';
-import { IMX_ADDRESS_ZKEVM, PRODUCTION_CHAIN_ID_NETWORK_MAP, SANDBOX_CHAIN_ID_NETWORK_MAP } from '../lib';
+import { IMX_ADDRESS_ZKEVM, PRODUCTION_CHAIN_ID_NETWORK_MAP, SANDBOX_CHAIN_ID_NETWORK_MAP } from '../env';
 
 let windowSpy: any;
 const providerMock = {
