+++ packages/checkout/sdk/src/smartCheckout/allowance/erc20.ts
@@ -3,7 +3,7 @@ import { BigNumber, Contract } from 'ethers';
 import { CheckoutError, CheckoutErrorType } from '../../errors';
 import { ItemRequirement, ItemType } from '../../types';
 import { Allowance, InsufficientERC20 } from './types';
-import { ERC20ABI } from '../../lib';
+import { ERC20ABI } from '../../env';
 
 // Gets the amount an address has allowed to be spent by the spender for the ERC20.
 export const getERC20Allowance = async (
