+++ packages/checkout/sdk/src/smartCheckout/allowance/erc721.ts
@@ -3,7 +3,7 @@ import { Contract } from 'ethers';
 import { CheckoutError, CheckoutErrorType } from '../../errors';
 import { ItemRequirement, ItemType } from '../../types';
 import { Allowance, InsufficientERC721 } from './types';
-import { ERC721ABI } from '../../lib';
+import { ERC721ABI } from '../../env';
 
 // Returns true if the spender address is approved for all ERC721s of this collection
 export const getERC721ApprovedForAll = async (
