+++ packages/checkout/sdk/src/smartCheckout/routing/bridge/bridgeRoute.test.ts
@@ -23,7 +23,7 @@ import { getBridgeFeeEstimate } from './getBridgeFeeEstimate';
 import { CheckoutErrorType } from '../../../errors';
 import { allowListCheckForBridge } from '../../allowList/allowListCheck';
 import { INDEXER_ETH_ROOT_CONTRACT_ADDRESS } from '../indexer/fetchL1Representation';
-import { DEFAULT_TOKEN_DECIMALS } from '../../../lib';
+import { DEFAULT_TOKEN_DECIMALS } from '../../../env';
 
 jest.mock('../../../gasEstimate');
 jest.mock('../../../instance');
