+++ packages/checkout/sdk/src/smartCheckout/routing/routingCalculator.test.ts
@@ -33,7 +33,7 @@ import { onRampRoute } from './onRamp';
 import { bridgeAndSwapRoute } from './bridgeAndSwap/bridgeAndSwapRoute';
 import { RoutingTokensAllowList } from '../allowList/types';
 import { INDEXER_ETH_ROOT_CONTRACT_ADDRESS } from './indexer/fetchL1Representation';
-import { IMX_ADDRESS_ZKEVM } from '../../lib';
+import { IMX_ADDRESS_ZKEVM } from '../../env';
 
 jest.mock('./tokenBalances');
 jest.mock('./bridge/bridgeRoute');
