+++ packages/checkout/sdk/src/widgets/load.test.ts
@@ -1,4 +1,4 @@
-import { SDK_VERSION_MARKER } from '../lib/env';
+import { SDK_VERSION_MARKER } from '../env/env';
 import { loadUnresolved } from './load';
 
 describe('load', () => {
